chore: move all revenue streams files in named folder
